+++ src/main/java/com/github/michaeldoyle/istat4j/model/TelemetryNetwork.java
@@ -1,5 +1,5 @@
 /* 
- * Copyright 2013 Michael J Doyle
+ * Copyright 2013-2014 Michael J Doyle
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,16 +20,16 @@ import java.util.Date;
 
 public class TelemetryNetwork {
 	
-	private long id;
+	private BigInteger id;
    	private BigInteger downloadBytes;
 	private BigInteger uploadBytes;
 	private Date time;
 	
-	public long getId() {
+	public BigInteger getId() {
 		return id;
 	}
 	
-	public void setId(long id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
